Build a camera
A camera is an entity that has a Transform
component that determines where it is placed in
the scene and the function that maps points onto
its film.

It also has an Image component that will contain
the captured image, and a Name component to make
the entity easier to identify.