Spike into using dynamic dispatch with components
Entity Component Systems help to decouple state
from behaviour but there are often common sets of
behaviours that many systems depend on.

For example, testing a ray-sphere intersection is
needed for the ray tracer, but it could also be
needed for a ‘picker’ system that lets you click
on a pixel and select an object in the scene.

Therefore, sometimes it makes sense to share
behaviours between systems, but in a way that’s
modular so that systems can pick-and-choose which
bits of behaviour they need.

This branch is a quick spike into how to do this.
It’s messy - but it works. It does this:

1) Add a trait for the behaviour we’re interested
in sharing. In this case, that’s ‘Shape’, but it
could be more granular and be ‘Intersectable’.

2) Add a component that owns a boxed trait object.
This requires the trait be Send + Sync so it can
work with specs’ parallelism.

Note: We also require the ‘static lifetime, since
the boxed value must be owned, rather than a
reference to something else. See:
https://stackoverflow.com/questions/26212397/references-to-traits-in-structs

3) Use the component containing the trait object
instead of the concrete values in the system(s)